Split reporting and counting into Authify.Stats context
Extract system analytics, counting, and organization-with-stats functions
from the Accounts context into a dedicated Authify.Stats context module.
Update all callers to use the new context.
