+++ lib/authify_web/controllers/dashboard_controller.ex
@@ -2,8 +2,8 @@ defmodule AuthifyWeb.DashboardController do
   use AuthifyWeb, :controller
 
   alias Authify.Accounts
-
   alias Authify.Invitations
+  alias Authify.Stats
 
   def index(conn, _params) do
     user = conn.assigns.current_user
@@ -11,7 +11,7 @@ defmodule AuthifyWeb.DashboardController do
 
     if organization.slug == "authify-global" do
       # Global admin dashboard - show system-wide data
-      system_stats = Accounts.get_system_stats()
+      system_stats = Stats.get_system_stats()
       invitation_stats = Invitations.get_invitation_stats()
 
       # Preload emails for recent users
