+++ lib/authify_web/controllers/maintenance_controller.ex
@@ -2,8 +2,8 @@ defmodule AuthifyWeb.MaintenanceController do
   use AuthifyWeb, :controller
 
   alias Authify.Accounts
-
   alias Authify.Invitations
+  alias Authify.Stats
   alias Authify.Tasks
 
   # All actions require being in the global organization
@@ -91,8 +91,8 @@ defmodule AuthifyWeb.MaintenanceController do
   defp get_database_stats do
     %{
       total_records: %{
-        organizations: Accounts.count_organizations(),
-        users: Accounts.count_users(),
+        organizations: Stats.count_organizations(),
+        users: Stats.count_users(),
         invitations: Invitations.count_invitations()
       },
       storage_usage: %{
@@ -119,7 +119,7 @@ defmodule AuthifyWeb.MaintenanceController do
       expired_invitations: Invitations.count_expired_invitations(),
       cleanable_invitations: Invitations.count_cleanable_invitations(),
       inactive_organizations_90d:
-        Accounts.count_inactive_organizations_since(DateTime.add(now, -90, :day)),
+        Stats.count_inactive_organizations_since(DateTime.add(now, -90, :day)),
       orphaned_sessions: get_orphaned_sessions_count(),
       temp_files: get_temp_files_count()
     }
